Add formatting/linting tools to repo.
These tools will automatically check and apply formatting rules.

This commit just includes the addition of the tools. We should do the
actual reformting separately after all the outstanding features have
been merged. For a preview, you can run

     poetry run pre-commit run --all-files

to run these hooks manually, but be aware it will modify almost every
source file in the repository.

Once we have made the code changes, we can add this as a git commit,
so that it will run automatically when you commit.

All the tools (black, ruff, mypy, etc.) are configurable. I have
largely kept the defaults for now, but if people have strong
feelings (for example on line length), we should decide now.