Change AWS credentials to use default provider chain.
This configures the Textract tests to always use the default provider
chain rather than passing in an explicit profile_name. This is
necessarily to run in Github actions. To run locally with a profile,
you can set the profile to be the default in your .aws/config file.

Also fixes some imports.