Add notebook tests and utilities to strip output from notebooks.
This commit pulls in nbmake as a dependency so that we can
automatically test the code in Jupyter notebooks for regressions. You
can run tests using

    poetry run pytest --nbmake notebooks/*.ipynb

This will excecute each cell of the notebooks and ensure they run
without error. It does not do any kind of corectness checking.

Notebooks typically cover end-to-end scenarios, so these tests are
effectively integration tests, and they require, for example, and
OpenAI API key and local OpenSearch instance. For this reason they are
not run automatically during pre-commit. Instead, I've added them as
another step in the integration test GitHub workflow.

These tests exposed some errors in tutorials.ipynb, which are also
fixed in this commit.

Finally this commit adds a pre-commit step that strips output from
Jupyter notebooks. This is just a convenience to avoid having to deal
with big git diffs that coming from changing output, which we don't
want to check-in anyway.