Parameter to help with debugging bounding box issues.
This is a minor thing, but I've found myself needing it twice in the
past couple of weeks, so it seems worth adding. If you initialize the
partitioner as `UnstructuredPdfPartitioner(retain_coordinates=True)`,
then the resulting elements will have a `coordinates` property with
the original Unstructured coordinates in addition to the bbox property
that we compute.