Update JSON scan to use Ray JSON reader.
In previous versions of Ray, the JSON reader did not retain the path
of the input file, which led us to do JSON parsing ourself. This has
been fixed in Ray 2.9.0, so we switch to using their JSON reader. This
has the byproduct of also adding support for line-delimited json
files (with the .jsonl extension).