Add some useful utility methods for DocSets.
This commit adds the `with_property`, `with_properties`, and
`map_elements`. This are primarily syntactic sugar around map
operations that I have found to be useful when writing complex
scripts. The `with_property` and `with_properties` methods simplify
adding properties to each Document in the in the DocSet. They are
loosely modeled after similar methods on Spark Datasets.

The `map_elements` method applies a function to each element in each
Document in the DocSet. This is a thin wrapper around the
Document-level `map` operation, but helps reduce boilerplate.