Remove mypy --explicit-package-bases flag and fix issues.
Since the refactor, I've noticed that the Github actions mypy check
was not catching all the mypy issues that I got when running `poetry
run mypy .` locally. I believe it is the explicit-package-bases flag.
Since we have __init__.py files in our main packages, we shouldn't
need it, and it seems to be messing up the module path discovery.