Update data model for table structure recognition. (#357)
* Update data model for table structure recognition.

This PR introduces a logical Table class that can be used to represent
the data in a document table in a format independent way. It is based
on representations used by TableTransformers and Textract, among
others. This commit just adds the basic data structure and some format
conversions and tests. Actual table understanding models will be added
in a follow-up PR.

This also includes some enhancements to the BoundingBox class so that
it can be used with a variety of table models.