Fix element creation in the Document class.
In the last PR, I added the create_element factory method so that we
got specific concrete element classes even after serialization.
However, I initially did it in the property getter, which leads to the
problem where mutations against an element don't get preserved. This
moves the logic to the constructor.