Refactor drawing code to support additional formats. (#409)
This commit introduces a new utility function to draw a collection of
boxes on a PIL image. This is used in both the DrawBoxes class and the
table drawing code. In general it is useful for testing new
OCR/detection methods, even when outside of DocSets.

This change should not impact existing public sycamore code.