Make it possible to pass in a schema for property extraction. (#481)
The ExtractProperties transform relies on a schema to guide the LLM on
what to extract. This schema can be extracted using an LLM, or it can
be added explicitly to each documetn in a "_schema" property.

I frequently want to be able to extract fields from an explicit schema
that I pass in, without having to update the properties map on each
document. This change just adds that ability in the form of arguments
to the OpenAIPropertyExtractor class. Example usage of how to extract
basic information about a scientific paper from a DocSet:

    schema = {
        "title": "string",
        "authors": "list[string]",
        "abstract": "string"
    }

    docs = context.read.binary(paths=[path], binary_format="pdf")\
                .partition(partitioner=SycamorePartitioner())\
                .extract_properties(OpenAIPropertyExtractor(llm, schema_name="PaperInfo", schema=schema))\
                .show(show_elements=False)