Fix long line issue in paritioner client
The default requests code can be quadratic in the data size if it finds a single long line
without a newline because it is doing the standard append to a string trick.  Switch over to
chunking based on raw data and explicitly handle splitting by lines for the early status output.