Return an empty table when table transformers don't find a table.
This was causing an assertion failure when the DETR model found a
table but table transformers didn't.

We still leave the assertion in for the case where the table
transformer model finds more than one table. It is not clear if this
is happening in practice, and the fix is more involved, so we defer it.