Fix materialize to work even if it is re-executed on the same documents.
The groupby_count operator when run via ray results in re-execution of
previous transforms.  The prior version of materialize meant this was an
error because it would detect duplicate files being written. The updated
code fixes this by including a sha checksum of the data in the default
filename.  This change means that perfect duplicates will be dropped by
materialize.

In order to preserve the safety properties, added a prepare operation on
Nodes so that materialize nodes can clear their paths and then check that
it was clear to make sure duplicate directories aren't being specified.