Reverts README change to restore poetry build. (#820)
A previous change attempted to avoid our README files getting out of
sync by referencing the README at the root of the repo using a
relative path in the library pyproject.toml. This unfortunately breaks
poetry build. This commit switches back to using the README in the lib
directory, and adds a github action step to copy the README during the
release process to hopefully avoid drift that way.