Add option to extract line-based bounding boxes from pdfminer.
We have been using pdfminer's layout detection to group text into
boxes. This can cause issues, especially with table extraction, when
the boxes don't line up with cells or what we detect with the DETR
model. This change adds support for an object_type parameter to the
PdfMinerExtractor that can be set to "boxes" (the current behavior),
or "lines", which groups characters into lines, but does not group
them further.

To avoid an explosion of options, we introduce a
"text_extractor_options" dict as a paramter, and refactor the
TextExtractor class hierarchy a bit to support it.