Add source_mode parameter to AutoMaterialize.
By passing in source_mode=MaterializeSourceMode.USE_STORED, this makes
it possible to avoid re-execution of intermediate plans when using
AutoMaterialize. The primary change is to avoid cleaning up the
materialized data at the beginning of every execution when the
source_mode is set to USE_STORED.

Also includes a utility for printing plans, which I found helpful when
debugging.