Add support for selecting pdf pages in Sycamore.
This copies functionality present in the Aryn Partitioning Service for
selecting specific pages or ranges of pages from pdfs. When applied to
a pdf Document, this will both update the binary_representation to
contain a valid pdf of only the selected pages, as well as retain only
the elements pertaining to those pages.