Support SummarizeImages with non-OpenAI LLMs. (#941)
This commit adds a generic LLMImageSummarizer that takes an LLM
instance. In order to account for differences in image formats
accepted by the LLMs, I also added a `format_image` method to the LLM
interface. Finally, I fixed a serialization issue in the Bedrock class
by not serializing the boto client.