Couple of minor fixes and tweaks to the table merger.
1. I observed the LLM sometimes having issues distinguishing where one
table ended and the next started, so I added headers - either ELEMENT
1/ELEMENT 2 or DOCUMENT 1/DOCUMENT 2 in the llm_query transform.

2. Added a default implementation for preprocess_element and
postprocess_element in the ElementMerger base class. This just reduces
code in some simple sub-classes.

3. Added code to update the table_continuation to be True or False.
This is because I ended up getting better results when asking the LLM
to explain its reasoning in some cases, but I didn't want all of the
output carried along in the element.

4. Minor fix to address an invalid escape character warning in a comment.