Fix OpenAI __reduce__ methods to handle Azure deployment names.
There was a bug in how we pickled OpenAI class instances specifically with
Azure. Azure OpenAI is a bit weird in that the "deployment name" of a model
can be different than the normal OpenAI name. This makes it impossible to use
the standard models from the OpenAIModels Enum. Instead, we typically created
separate OpenAIModel instances. The problem was that when we pickled the
OpenAI object, it saved just the model name, and then tried to recreate the
model option by using the Enum. This change pickles the entire
OpenAIModel class instead of just the name. Also adds a test that
demonstrates the bug (and fix). The other tests were passing because they
called the OpenAI methods directly and didn't pickle the object. The object
will be pickled in basically any Sycamore job using ray.