Add extract_image_format option. (#1230)
This can be set to any image format that PIL supports, and it controls the format an image is encoded by the partitioner when extract_images is True. The default
is PPM, which is the current behavior, but PPMs are not displayable on the console, so we provide an alternative option.