Fix fast notebook tests (#1316)
Previously the _infer_prompt method would call asyncio.run to start up an event loop in the case that an LLM was in async mode. This causes issues in
environments like Jupyter notebooks that already have a running event loop. To avoid conflicts, we modify this to start a new thread and run the new event loop there.

Also switches from GPT_3_5_TURBO_INSTRUCT to GPT_4O_MINI as I was getting context length errors and 4o-mini is cheaper anyway.