Fix issue with nested_lookup utility. (#1320)
This method is used in some of the Jinja prompt formatters via Element::field_to_value. It previously assumed the object it was traversing was a dict and used the get method to look up values. In elements, this bypassed the @property getters, which could lead to incorrect values. I observed the text_representation of a TableElement being null, because the value is set to the csv representation of the table in the getter.