Drop support for Python 3.9 and add Python 3.13.
This change does three things:

- Drops support for Python 3.9 and removes it from our Github Actions tests.

- Adds support for Python 3.13. This required updating a few
dependencies (primarily scipy, matplotlib, and pandas).

- Adds sycamore to the default ray runtime_env. I am not sure why this showed
up for me with this change. Somehow upgrading from 2.43.0 to 2.47.1 (which I
did not do explicitly, but must have come when I regenerated the poetry.lock)
seems to expose a ModuleNotFound error on the workers. I believe this is the
correct fix, based on
https://docs.ray.io/en/latest/ray-core/handling-dependencies.html#library-development,
but can't explain what changed.