Initial implementation of SchemaV2. (#1404)
This enhances the Schema specification that we intend to use for property
extraction. While I have made an effort to support backwards compatibility in
the sense of deserializing old schemas, I have put the new functionality under
SchemaV2 for the time being to avoid breaking anything. Eventually I will
merge them and move to using Schema.