Make file extension comparison case insensitive in local mode. (#1468)
When you specify a binary_format to BinaryScan, it restricts the files to only those of the specified format. Previously there was an inconsistency in behavior between Ray and local modes. In Ray mode, the extensions were specified in a case-insensitive way, whereas in local mode they were case sensitive.

This commit makes both case insensitive.