Initial Iceberg writer.
This commit contains an Iceberg Write node that writes properties to an
Iceberg table using the PyIceberg library.

The implementation differs somewhat for local mode and Ray mode. In local
mode, we convert the DocSet into a pyarrow table and use pyiceberg's native
arrow support for writing. In Ray mode, we still leverage pyarrow, but we use
it to convert the Docset into a Ray Dataset of the properties and then use the
Ray Data Iceberg writer directly.