Update schema deserialization fallback behavior. (#1497)
If we can't deserialize a schema as SchemaV2, we attempt to deserialize it as a Schema and then convert. This means that in the case of a malformed schema, we would always raise the Exception from the fallback, which can lead to confusing error messages. This change updates this behavior to raise the original exception in case the fallback fails.