Attempt to fix flaky unit test due to Ray Dataset ordering.
These tests have been flaky in Github recently, apparently because they were
making assumptions about the order in which records were returned from a Ray
dataset. It is possible that this is due to a recent Ray version update,
though I could not identify any obvious cause or change in contract. This
commit updates the tests to make the assertions slightly less strict.