Don't default to all elements for LLM-based attribution.
The text-based attribution strategy has a heuristic where attribution is
initialized to the set of elements passed to the llm in a given call. This was
being applied with the LLMAttributionStrategy, which was leading to cases
where the attribution was being set to all of the elements instead of None.
This is not what we want. This PR just makes the default a method in the
AttributionStrategy and does the right thing in the LLM strategy.